feat: add checks is user an admin

+++ routes/profiles.routes.js
@@ -7,6 +7,13 @@ router.get(
     '/',
     async (req, res) => {
       try {
+        const {isUserAdmin} = req.body;
+        if (!isUserAdmin) {
+          return res.status(401).json({
+            message: 'Вы не админ',
+          });
+        }
+
         const profiles = await Profile.find();
         res.json(profiles);
       } catch (error) {
@@ -36,6 +43,13 @@ router.delete(
     '/delete/:id',
     async (req, res) => {
       try {
+        const {isUserAdmin} = req.body;
+        if (!isUserAdmin) {
+          return res.status(401).json({
+            message: 'Вы не админ',
+          });
+        }
+
         const profileId = req.params.id;
 
         await Profile.deleteOne({_id: profileId});
