+++ routes/users.routes.js
@@ -7,6 +7,13 @@ router.get(
     '/',
     async (req, res) => {
       try {
+        const {isUserAdmin} = req.body;
+        if (!isUserAdmin) {
+          return res.status(401).json({
+            message: 'Вы не админ',
+          });
+        }
+
         const users = await User.find();
         res.json(users);
       } catch (error) {
@@ -22,6 +29,13 @@ router.delete(
     '/delete/:email',
     async (req, res) => {
       try {
+        const {isUserAdmin} = req.body;
+        if (!isUserAdmin) {
+          return res.status(401).json({
+            message: 'Вы не админ',
+          });
+        }
+
         const userEmail = req.params.email;
         console.log(userEmail);
         await User.deleteOne({email: userEmail});
